organized the SettingsTabs folders
