+++ client/src/components/Nav/SettingsTabs/General/AutoScrollSwitch.spec.tsx
