+++ client/src/components/Nav/SettingsTabs/General/AutoScrollSwitch.tsx
