+++ client/src/components/Nav/SettingsTabs/index.ts
@@ -1,5 +1,5 @@
-export { default as General } from './General';
-export { ClearChatsButton } from './General';
-export { default as Data } from './Data';
-export { RevokeKeysButton } from './Data';
-export { default as Account } from './Account';
+export { default as General } from './General/General';
+export { ClearChatsButton } from './General/General';
+export { default as Data } from './Data/Data';
+export { RevokeKeysButton } from './Data/Data';
+export { default as Account } from './Account/Account';
